Steam Release

+++ Source/Scenes/Cipher/Components/CipherLock.cpp
@@ -19,8 +19,8 @@
 
 using namespace cocos2d;
 
-const int CipherLock::PinSpacing = 48.0f;
-const int CipherLock::PinUnlockDistance = 104.0f;
+const float CipherLock::PinSpacing = 48.0f;
+const float CipherLock::PinUnlockDistance = 104.0f;
 
 CipherLock* CipherLock::create()
 {
@@ -103,7 +103,7 @@ void CipherLock::initializePositions()
 	this->stoppingBlock->setPosition(Vec2(-16.0f, 96.0f));
 	this->pinboardFront->setPosition(Vec2(-40.0f, 0.0f));
 
-	for (int index = 0; (index < this->cipherPins.size() && index < this->cipherPinholes.size()); index++)
+	for (int index = 0; (index < int(this->cipherPins.size()) && index < int(this->cipherPinholes.size())); index++)
 	{
 		this->cipherPinholes[index]->setPosition(Vec2(0.0f, float(index) * -CipherLock::PinSpacing));
 		this->cipherPins[index]->setPosition(Vec2(-24.0f, float(index) * -CipherLock::PinSpacing));
@@ -122,7 +122,7 @@ void CipherLock::initializeListeners()
 		{
 			this->hasAnyPinFailed |= !(args->success);
 			
-			if (args->cipherIndex < cipherPins.size())
+			if (args->cipherIndex < int(cipherPins.size()))
 			{
 				Sprite* pinhole = this->cipherPinholes[args->cipherIndex];
 				Sprite* pin = this->cipherPins[args->cipherIndex];
